+++ src/cli/mapshaper-option-validation.js
@@ -54,20 +54,25 @@ function validateProjOpts(cmd) {
   var _ = cmd._,
       proj4 = [];
 
-  // separate proj4 options
-  _ = _.filter(function(arg) {
-    if (/^\+[a-z]/i.test(arg)) {
-      proj4.push(arg);
-      return false;
-    }
-    return true;
-  });
-
-  if (proj4.length > 0) {
-    cmd.options.crs = proj4.join(' ');
-  } else if (_.length > 0) {
-    cmd.options.crs = _.shift();
+  if (_.length > 0 && !cmd.options.crs) {
+    cmd.options.crs = _.join(' ');
+    _ = [];
   }
+
+  // separate proj4 options
+  // _ = _.filter(function(arg) {
+  //   if (/^\+[a-z]/i.test(arg)) {
+  //     proj4.push(arg);
+  //     return false;
+  //   }
+  //   return true;
+  // });
+
+  // if (proj4.length > 0) {
+  //   cmd.options.crs = proj4.join(' ');
+  // } else if (_.length > 0) {
+  //   cmd.options.crs = _.shift();
+  // }
 
   if (_.length > 0) {
     error("Received one or more unexpected parameters: " + _.join(', '));